% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\name{alldata}
\alias{alldata}
\title{Reading is alldata}
\usage{
alldata(dir, pattern = NULL, header = T, sep = ",", skip = 0,
  encoding = "unknown", stringsAsFactors = F, t.data = NULL,
  temp = NULL, sheetIndex = 1)
}
\arguments{
\item{dir}{is directory.}
}
\value{
alldata  #Sys.setenv( JAVA_HOME = 'C:\\Program Files\\Java\\jre1.8.0_151' )
}
\description{
Reading is alldata
}
