% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change Radian to degree.R
\name{by_angle}
\alias{by_angle}
\title{Change Radian to degree}
\usage{
by_angle(x)
}
\arguments{
\item{x}{is angle}
}
\value{
the radian conversion value of x
}
\description{
Change Radian to degree
}
\examples{
sin(by_ange(90))
}
