% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ganerate cofactor matrix or adjoint matrix.R
\name{cofactor}
\alias{cofactor}
\title{Ganerate matrix of cofactor from A.(or adjont of a matrix A)}
\usage{
cofactor(A, adj = F)
}
\arguments{
\item{A}{is numerix matrix}
}
\value{
cofactor
}
\description{
Ganerate matrix of cofactor from A.(or adjont of a matrix A)
}
\examples{
A<-matrix(c(2,-1,1,1,2,0,-3,-2,4),nrow=3,ncol=3)
cofactor(A)
cofactor(A,adj=T)
}
