% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-fold Cross Validation.R
\name{cv.k_fold}
\alias{cv.k_fold}
\title{K-fold cross Validation}
\usage{
cv.k_fold(data, k, tr = list(), te = list(), print = F)
}
\arguments{
\item{k=number}{of folder}
}
\value{
train data and test data of list form data.
}
\description{
K-fold cross Validation
}
\examples{
cv.k_fold(iris,4,print=T)
}
