% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing value processing.R
\name{na_rm}
\alias{na_rm}
\title{Missing value processing}
\usage{
na_rm(data, na.all = F)
}
\arguments{
\item{na.all}{Whether or not to remove at least one missing value}
}
\value{
Data without missing values
}
\description{
Missing value processing
}
\examples{
a<-c(NA,NA,NA,NA,1);b<-c(NA,NA,NA,1,2);c<-c(NA,NA,1,2,3);d<-c(NA,1,2,3,4)
(data<-data.frame(a,b,c,d))
na_rm(data)
na_rm(data,na.all=T)
}
