% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_arrows.R
\name{p_arrows}
\alias{p_arrows}
\title{arrows is overlay in plot}
\usage{
p_arrows(length = 0.25, angle = 30, code = 2, col = par("fg"),
  lty = par("lty"), lwd = par("lwd"))
}
\arguments{
\item{non}{parameters.}
}
\value{
overlab of plot
}
\description{
arrows is overlay in plot
}
\examples{
plot(1)
p_arrows()
}
