% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_legend.R
\name{p_legend}
\alias{p_legend}
\title{legend is overlay in plot}
\usage{
p_legend(legend, fill = NULL, col = par("col"), border = "black", lty,
  lwd, pch, angle = 45, density = NULL, bty = "o", bg = par("bg"),
  box.lwd = par("lwd"), box.lty = par("lty"), box.col = par("fg"),
  pt.bg = NA, cex = 1, pt.cex = cex, pt.lwd = lwd, xjust = 0,
  yjust = 1, x.intersp = 1, y.intersp = 1, adj = c(0, 0.5),
  text.width = NULL, text.col = par("col"), text.font = NULL,
  merge = do.lines && has.pch, trace = FALSE, plot = TRUE, ncol = 1,
  horiz = FALSE, title = NULL, inset = 0, xpd, title.col = text.col,
  title.adj = 0.5, seg.len = 2)
}
\arguments{
\item{non}{parameters.}
}
\value{
overlab of plot
}
\description{
legend is overlay in plot
}
\examples{
plot(1)
p_legend('x')
}
