% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package}
\alias{package}
\title{install & library both summit}
\usage{
package(package, character.only = F, quietly = F, lib.loc = NULL,
  warn.conflicts = T)
}
\arguments{
\item{package}{install & load}
}
\value{
library
}
\description{
install & library both summit
}
