% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_all.R
\name{shp_all}
\alias{shp_all}
\title{Read 'shp files' from 'folder' in 'working directory' and save.}
\usage{
shp_all(wd, a = list())
}
\arguments{
\item{wd}{is working directory}
}
\value{
print 'shp file' as 'list'
}
\description{
Read 'shp files' from 'folder' in 'working directory' and save.
}
\examples{
wd<-'C:/Users/changje/Desktop/shp'
b<-shp_all(wd)
plot(b[[1]])
package(ggmap)
kor<-fortify(b[[1]])
ggplot(data=kor,aes(x=long,y=lat,group=group, colour = id))+geom_polygon(fill = "#FFFFFF")
}
