% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular matrix.R
\name{tri}
\alias{tri}
\title{matrix to a triangular matrix}
\usage{
tri(x, diag = TRUE, upper = T)
}
\arguments{
\item{x}{is numerix matrix form}
}
\value{
upper(lower) triangular matrix
}
\description{
matrix to a triangular matrix
}
\examples{
x<-matrix(1,nrow=4,ncol=3)
tri(x)
}
