% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfactor.R
\name{unfactor}
\alias{unfactor}
\title{change 'factor type data' to 'character type data'}
\usage{
unfactor(x)
}
\arguments{
\item{x}{a factor type data.}
}
\value{
character type data.
}
\description{
change 'factor type data' to 'character type data'
}
\examples{
unfactor(iris[,4])
unfactor(iris[,5])
}
